A natural language processing project
The project aimed to examine the similarities of the languages ​​of the former Soviet Union countries. Language is very important for science and art. Each language in the world is unique in its richness. Throughout history, languages ​​have been influenced by each other over time. It was inevitable that the Caucasian, Turkish, Slavic and Baltic nations, which were once under the same roof, were in this situation. In many parts of the world, small linguistic differences from one town to another create a completely different language within a half-hour drive. For its breathtaking complexity and diversity, human language is like a colorful tapestry stretching across time and space.